...
git-svn-id: https://repo.acin.tuwien.ac.at/v4r/trunk/ros/hobbit_hydro_src@9656 a44b1c4d-ba22-49c9-afbd-3e401e1c4b3e

+++ emergency_detector/src/process.h
@@ -24,8 +24,8 @@ extern unsigned int tempTimestamp;
 
 extern  int maximumFrameDifferenceForTemperatureToBeRelevant;
 
-extern  int minimumAllowedHolePercentage;
-extern  int maximumAllowedHolePercentage;
+extern  double minimumAllowedHolePercentage;
+extern  double maximumAllowedHolePercentage;
 
 extern double minHumanTemperature;
 extern double maxHumanTemperature;
